--- cli/src/deps.ts
@@ -1,7 +1,0 @@
-export type {
-  LoggerAction,
-  LoggerInterface,
-  TestModule,
-} from "../../core/dist/mod.js";
-
-export { startRun, cancelRun } from "../../core/dist/mod.js";
