+++ tests/src/mod.ts
@@ -1,38 +1,8 @@
 import * as FailTests from "./test_fail.test.js";
 import * as PassTests from "./test_pass.test.js";
 
-import type {
-  LoggerAction,
-  LoggerInterface,
-  TestModule,
-} from "../../core/dist/mod.js";
-
 import { startRun } from "../../core/dist/mod.js";
-
-class TestLogger implements LoggerInterface {
-  cancelled: boolean;
-  has_failed: boolean = false;
-
-  log(_testModule: TestModule[], action: LoggerAction) {
-    if (hasTestFailed(action)) {
-      this.has_failed = true;
-    }
-  }
-
-  get result(): boolean {
-    return this.has_failed;
-  }
-}
-
-function hasTestFailed(action: LoggerAction) {
-  if ("end_test" !== action.type) return false;
-
-  if (action.assertions === undefined) return false;
-  if (Array.isArray(action.assertions) && action.assertions.length === 0)
-    return false;
-
-  return true;
-}
+import { TestLogger } from "./test_logger.js";
 
 // Test pass and fail behavior
